* feat(migrate): add MigrateReviewSign component * refactor(pages/migrate/review.vue): use MigrateReviewSign component * refactor(pages/migrate/sign.vue): use MigrateReviewSign component
